Add logging to the rpc package.
I have added a few info log statements for general API invocations with little data.
At the FINE level each request/response will be printed.
At the FINER level each method invocation with parameters and results are printed.
At the FINEST level the decoded body both as json and as bytes is printed.

When returning an API error it is printed at the WARNING level.
When returning a ApiConfigError it is printed at the SEVERE level (it will not run).

R=sgjesse@google.com, kustermann@google.com, lukechurch@google.com
BUG=

Review URL: https://codereview.chromium.org/951363002