Support for constructors taking arguments for message classes only used as responses.
In most cases the messages used entirely as response classes would prefer to have a constructor taking the result values as arguments. There is no reason why we cannot allow that as we don't instantiate these classes using reflection. However if the class is used as a request we still require the no-args default constructor.

R=kustermann@google.com, sgjesse@google.com, lukechurch@google.com
BUG=

Review URL: https://codereview.chromium.org/1130453002