Add cookies, response status code, and response headers to the invocation context.
This allows the implementor of the api method to affect the response and to see what cookies were passed in the request.
Also changed the default status code of a successful request to be NO_CONTENT if not body is returned in the response.
Finally fixed a bug when failing with a FormatException when decoding a request body and not getting 'unexpected end of input'. In that case we just tried invoking the method instead of failing.
R=ricow@google.com, kustermann@google.com, lrn@google.com, sgjesse@google.com
BUG=

Review URL: https://codereview.chromium.org/1133483005